# lomix-next/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lomix Mobile API Documentation",
    "version": "1.0.0",
    "description": "API documentation for Lomix Mobile Application"
  },
  "servers": [
    {
      "url": "https://lomix-be-i2rg.vercel.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local server (Development)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/mobile/room/join": {
      "post": {
        "summary": "Odaya Katıl",
        "tags": [
          "Mobile Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roomId"
                ],
                "properties": {
                  "roomId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Odaya başarıyla katıldınız"
          },
          "403": {
            "description": "Seviye yetersiz veya oda aktif değil"
          },
          "404": {
            "description": "Oda bulunamadı"
          }
        }
      }
    },
...
    "/api/mobile/room/member-request": {
      "post": {
        "summary": "Odaya üyelik isteği gönder",
        "tags": [
          "Mobile Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roomId"
                ],
                "properties": {
                  "roomId": {
                    "type": "string",
                    "description": "Oda ID (roomId alanı)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "İstek gönderildi"
          },
          "409": {
            "description": "Zaten istek gönderilmiş veya zaten üye"
          }
        }
      },
      "put": {
        "summary": "Üyelik isteğini kabul et / reddet (oda sahibi/admin)",
        "tags": [
          "Mobile Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "requestId",
                  "action"
                ],
                "properties": {
                  "requestId": {
                    "type": "integer"
                  },
                  "action": {
                    "type": "string",
                    "enum": [
                      "accept",
                      "reject"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "İşlem tamamlandı"
          }
        }
      }
    },
    "/api/mobile/room/member-requests": {
      "get": {
        "summary": "Odaya gelen üyelik isteklerini listele (oda sahibi/admin)",
        "tags": [
          "Mobile Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "roomId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Oda ID (roomId alanı)"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "accepted",
                "rejected"
              ]
            },
            "description": "Filtre (varsayılan pending)"
          }
        ],
        "responses": {
          "200": {
            "description": "İstekler listelendi"
          }
        }
      }
    },
...
  },
...
}